Fix callback implementation
